requirements.txt
pandas
langchain
langchain_community
sentence-transformers
torch
pypdf
qdrant-client